Deleted eda_visualization and added the graph to a ny file in frontend called graphs and added the nearbystops function to plot_maps in frontend.
